requirements.txt
streamlit>=1.0.0
langchain>=0.0.200
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-pinecone>=0.1.0
pinecone-client>=2.2.1
openai>=1.1.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.0
prometheus-client>=0.16.0
pydantic>=2.0.0
langsmith>=0.1.17
backoff>=2.0.0
pypdf>=3.0.0
typing-extensions>=4.5.0
boto3>=1.26.0
tiktoken>=0.5.0 